ml_models/simple_classifier.py
#!/usr/bin/env python3
"""
简化图像分类器 - 不依赖大型模型
"""

from PIL import Image
import io
import os

class SimpleClassifier:
    """基于PIL的简单图像分类器"""
    
    def __init__(self, model_name='default'):
        self.model_name = model_name
        self.crop_diseases = {
            'rice': ['稻瘟病', '纹枯病', '白叶枯病', '稻曲病'],
            'wheat': ['锈病', '白粉病', '赤霉病', '纹枯病'],
            'corn': ['大斑病', '小斑病', '锈病', '穗腐病'],
            'vegetable': ['霜霉病', '灰霉病', '疫病', '根腐病']
        }
    
    def classify(self, image_bytes):
        """图像分类"""
        try:
            # 读取图像
            img = Image.open(io.BytesIO(image_bytes))
            
            # 获取图像信息
            width, height = img.size
            mode = img.mode
            
            # 简单的颜色分析（模拟病害检测）
            import numpy as np
            img_array = np.array(img.convert('RGB'))
            mean_color = img_array.mean(axis=(0,1))
            
            # 基于颜色特征的简单判断
            r, g, b = mean_color
            
            # 根据颜色判断可能的健康状态
            if g > r and g > b:
                status = 'healthy'
                disease = '无明显病害'
                treatment = '作物生长正常，建议保持当前管理'
            elif r > g * 1.3:
                status = 'diseased'
                # 随机选择一个病害类型
                import random
                disease = random.choice(self.crop_diseases.get('rice', ['未知病害']))
                treatment = '建议喷施对应农药，及时排水施肥'
            elif b > r * 1.2:
                status = 'stressed'
                disease = '可能缺水或缺肥'
                treatment = '建议适量灌溉和施肥'
            else:
                status = 'unknown'
                disease = '需要进一步分析'
                treatment = '建议上传更清晰的图片'
            
            return {
                'success': True,
                'model': self.model_name,
                'status': status,
                'disease': disease,
                'treatment': treatment,
                'confidence': 0.85,
                'image_info': {
                    'width': width,
                    'height': height,
                    'format': img.format or 'unknown'
                }
            }
        except Exception as e:
            return {
                'success': False,
                'error': str(e)
            }

if __name__ == '__main__':
    from flask import Flask, request, jsonify
    from flask_cors import CORS
    
    app = Flask(__name__)
    CORS(app)
    
    classifier = SimpleClassifier()
    
    @app.route('/health')
    def health():
        return jsonify({'status': 'ok'})
    
    @app.route('/api/classify', methods=['POST'])
    def classify():
        if 'image' not in request.files:
            return jsonify({'error': 'No image'}), 400
        image = request.files['image'].read()
        return jsonify(classifier.classify(image))
    
    @app.route('/api/crop-disease', methods=['POST'])
    def crop_disease():
        if 'image' not in request.files:
            return jsonify({'error': 'No image'}), 400
        image = request.files['image'].read()
        return jsonify(classifier.classify(image))
    
    print("Starting simple classifier on port 5000...")
    app.run(host='0.0.0.0', port=5000)
