ml_models/requirements.txt
torch
torchvision
timm
ultralytics
flask
flask-cors
pillow
numpy
